Refactor DKG into fedimint-server
