+++ fedimint-server/Cargo.toml
@@ -13,6 +13,7 @@ name = "fedimint_server"
 path = "src/lib.rs"
 
 [dependencies]
+aead = { path = "../crypto/aead" }
 anyhow = "1.0.66"
 async-trait = "0.1.64"
 bincode = "1.3.1"
@@ -23,7 +24,6 @@ hbbft = { git = "https://github.com/jkitman/hbbft", branch = "upgrade-threshold-
 futures = "0.3.24"
 impl-tools = "0.6.1"
 itertools = "0.10.5"
-jsonrpsee = { version = "0.16.2", features = ["server"] }
 fedimint-api = { path = "../fedimint-api" }
 fedimint-core = { path = "../fedimint-core" }
 rand = "0.8"
@@ -37,13 +37,14 @@ strum = "0.24"
 strum_macros = "0.24"
 tbs = { path = "../crypto/tbs" }
 thiserror = "1.0.37"
+tracing ="0.1.37"
+url = { version = "2.3.1", features = ["serde"] }
+threshold_crypto = { git = "https://github.com/jkitman/threshold_crypto", branch = "upgrade-threshold-crypto-libs" }
+jsonrpsee = { version = "0.16.2", features = ["server"] }
 tokio = { version = "1.25.0", features = ["full"] }
 tokio-stream = "0.1.11"
 tokio-rustls = "0.23.4"
 tokio-util = { version = "0.7.4", features = [ "codec" ] }
-tracing ="0.1.37"
-url = { version = "2.3.1", features = ["serde"] }
-threshold_crypto = { git = "https://github.com/jkitman/threshold_crypto", branch = "upgrade-threshold-crypto-libs" }
 
 [dev-dependencies]
 tracing-subscriber = { version = "0.3.16", features = [ "env-filter" ] }
