+++ fedimintd/Cargo.toml
@@ -36,9 +36,9 @@ fedimint-api = { path = "../fedimint-api" }
 fedimint-core = { path = "../fedimint-core" }
 fedimint-rocksdb = { path = "../fedimint-rocksdb" }
 fedimint-server = { path = "../fedimint-server" }
-fedimint-wallet = { path = "../modules/fedimint-wallet", features = ["native"] }
-fedimint-mint= { path = "../modules/fedimint-mint" }
-fedimint-ln = { path = "../modules/fedimint-ln" }
+fedimint-wallet = { path = "../modules/fedimint-wallet", features = ["native", "server"] }
+fedimint-mint= { path = "../modules/fedimint-mint", features = ["server"] }
+fedimint-ln = { path = "../modules/fedimint-ln", features = ["server"] }
 opentelemetry = { version = "0.18.0", optional = true }
 opentelemetry-jaeger = { version = "0.17.0", optional = true }
 rand = "0.8"
