+++ fedimintd/src/ui.rs
@@ -19,6 +19,10 @@ use fedimint_api::task::TaskGroup;
 use fedimint_api::Amount;
 use fedimint_core::api::WsClientConnectInfo;
 use fedimint_core::util::SanitizedUrl;
+use fedimint_server::config::io::{
+    create_cert, encrypted_json_write, parse_peer_params, run_dkg, write_nonprivate_configs,
+    CONSENSUS_CONFIG, JSON_EXT, PRIVATE_CONFIG, SALT_FILE, TLS_PK,
+};
 use http::StatusCode;
 use qrcode_generator::QrCodeEcc;
 use serde::Deserialize;
@@ -29,11 +33,7 @@ use tokio_rustls::rustls;
 use tracing::{debug, error};
 use url::Url;
 
-use crate::distributedgen::{create_cert, parse_peer_params, run_dkg};
-use crate::{
-    configure_modules, encrypted_json_write, write_nonprivate_configs, CONSENSUS_CONFIG, JSON_EXT,
-    PRIVATE_CONFIG, SALT_FILE, TLS_PK,
-};
+use crate::{configure_modules, module_registry, CODE_VERSION};
 
 #[derive(Deserialize, Debug, Clone)]
 #[allow(dead_code)]
@@ -182,7 +182,9 @@ async fn post_guardians(
                 connection_strings,
                 rustls::PrivateKey(pk_bytes),
                 &mut dkg_task_group,
+                CODE_VERSION,
                 configure_modules(max_denomination, params.network, params.finality_delay),
+                module_registry(),
             )
             .await;
