+++ modules/fedimint-wallet/Cargo.toml
@@ -14,6 +14,8 @@ path = "src/lib.rs"
 [features]
 native = ["tokio", "fedimint-bitcoind/bitcoincore-rpc"]
 default = []
+# FIXME: Currently required because the client depends on the server modules
+server = ["fedimint-server"]
 
 [dependencies]
 anyhow = "1.0.66"
@@ -35,6 +37,7 @@ tokio = { version = "1.25.0", features = ["sync"], optional = true }
 tracing ="0.1.37"
 url = "2.3.1"
 validator = { version = "0.16", features = ["derive"] }
+fedimint-server = { path = "../../fedimint-server", optional = true  }
 
 [dev-dependencies]
 test-log = { version = "0.2", features = [ "trace" ], default-features = false }
